+++ app/pages/capitalallowances/zeroEmissionGoodsVehicle/ZeroEmissionGoodsVehiclePage.scala
@@ -20,6 +20,7 @@ import controllers.journeys.capitalallowances.zeroEmissionGoodsVehicle.routes
 import models.NormalMode
 import models.common._
 import models.database.UserAnswers
+import pages.redirectOnBoolean
 import play.api.mvc.Call
 import queries.Settable
 
@@ -39,6 +40,7 @@ object ZeroEmissionGoodsVehiclePage extends ZegvBasePage[Boolean] {
 
   override def nextPageInNormalMode(userAnswers: UserAnswers, businessId: BusinessId, taxYear: TaxYear): Call =
     redirectOnBoolean(
+      this,
       userAnswers,
       businessId,
       onTrue = routes.ZegvAllowanceController.onPageLoad(taxYear, businessId, NormalMode),
